[#95] DynamicSize 구현

+++ EUMMEYO/Extension/CGDouble+.swift
@@ -1,5 +1,5 @@
 //
-//  CGFloat+.swift
+//  CGDouble+.swift
 //  EUMMEYO
 //
 //  Created by 김동현 on 4/19/25.
@@ -7,10 +7,8 @@
 
 import Foundation
 
-extension CGFloat {
+extension Double {
     var scaled: CGFloat {
-        return DynamicSize.scaledSize(self)
+        return DynamicSize.scaledSize(CGFloat(self))
     }
 }
-
-
